backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4

# Payments
stripe==7.11.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

# Templates
jinja2==3.1.3

# HTTP Client
httpx>=0.24,<0.26

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
